Fixed spelling in higlass_view_config schema and inserts
One of the "options" in the viewconf was plusStandColor rather than
plusStrandColor - corrected this in schema and inserts.

Note: in some inserts, under options, there were both plusStandColor
and plusStrandColor - usually with the former blue and the latter black.
In these cases I just deleted the former and kept the latter, which
remained black. In cases with only the former I corrected the spelling
of the property name.